[feat] 코드 실행시간 자료형 int로 변경

+++ app/judge_core.py
@@ -92,7 +92,7 @@ def judge_submission(req: SubmissionRequest) -> SubmissionResponse:
                 end_time = time.perf_counter()
                 usage = resource.getrusage(resource.RUSAGE_CHILDREN)
                 memory_used_kb = usage.ru_maxrss
-                time_taken_ms = (end_time - start_time) * 1000
+                time_taken_ms = int((end_time - start_time) * 1000)
 
             except subprocess.TimeoutExpired:
                 logger.warning("Time limit exceeded for test case: %s", input_path.name)
